Share process types in dprint_core
