+++ crates/core/src/process/mod.rs
@@ -1,3 +1,5 @@
 mod stdin_out_reader_writer;
+mod shared_types;
 
 pub use stdin_out_reader_writer::*;
+pub use shared_types::*;
