+++ crates/core/src/process/shared_types.rs
@@ -1,0 +1,71 @@
+/// The process plugin schema version.
+pub const PLUGIN_SCHEMA_VERSION: u32 = 1;
+
+/// Kinds of messages that process plugins must handle.
+#[derive(Debug)]
+pub enum MessageKind {
+    GetPluginSchemaVersion = 0,
+    GetPluginInfo = 1,
+    GetLicenseText = 2,
+    GetResolvedConfig = 3,
+    SetGlobalConfig = 4,
+    SetPluginConfig = 5,
+    GetConfigDiagnostics = 6,
+    /// Returns a format result part, then a file text part.
+    FormatText = 7,
+}
+
+// todo: generate with a macro
+impl From<u32> for MessageKind {
+    fn from(kind: u32) -> Self {
+        match kind {
+            0 => MessageKind::GetPluginSchemaVersion,
+            1 => MessageKind::GetPluginInfo,
+            2 => MessageKind::GetLicenseText,
+            3 => MessageKind::GetResolvedConfig,
+            4 => MessageKind::SetGlobalConfig,
+            5 => MessageKind::SetPluginConfig,
+            6 => MessageKind::GetConfigDiagnostics,
+            7 => MessageKind::FormatText,
+            _ => unreachable!("Unexpected message kind: {}", kind),
+        }
+    }
+}
+
+/// The kinds of responses.
+#[derive(Debug)]
+pub enum ResponseKind {
+    Success = 0,
+    Error = 1,
+}
+
+// todo: generate with a macro
+impl From<u32> for ResponseKind {
+    fn from(orig: u32) -> Self {
+        match orig {
+            0 => ResponseKind::Success,
+            1 => ResponseKind::Error,
+            _ => unreachable!("Unexpected response kind: {}", orig),
+        }
+    }
+}
+
+/// The kinds of format results.
+#[derive(Debug)]
+pub enum FormatResult {
+    NoChange = 0,
+    Change = 1,
+    Error = 2,
+}
+
+// todo: generate with a macro
+impl From<u32> for FormatResult {
+    fn from(orig: u32) -> Self {
+        match orig {
+            0 => FormatResult::NoChange,
+            1 => FormatResult::Change,
+            2 => FormatResult::Error,
+            _ => unreachable!("Unexpected format result: {}", orig),
+        }
+    }
+}
